// EscalafonFrontend/src/app/interfaces/acreditacion-lengua-nativa.ts
export interface AcreditacionLenguaNativa {
    /**
     * @type { number } - iAcredId
     */
    iAcredId?: number;

    /**
     * @type { number } - iLegId
     */
    iLegId?: number;

    /**
     * @type { string } - cAcredIdioma
     */
    cAcredIdioma?: string;

    /**
     * @type { string } - cAcredDominioEscritura
     */
    cAcredDominioEscritura?: string;

    /**
     * @type { string } - cAcredDominioOral
     */
    cAcredDominioOral?: string;

    /**
     * @type { number } - iAcredAnioIngreso
     */
    iAcredAnioIngreso?: number;

    /**
     * @type { number } - iAcredAnioEvaluacion
     */
    iAcredAnioEvaluacion?: number;

    /**
     * @type { number } - iAcredAnioVencimiento
     */
    iAcredAnioVencimiento?: number;

}